Lots of fixes
Made signals/slots use int, and cast to uin8_t to get Qt to stop complaining
Added some logging

+++ LEDManager.h
@@ -33,8 +33,8 @@ public:
 
 public slots:
 	void process();
-	void setBrightness(uint8_t);
-	void setRGB(uint8_t, uint8_t, uint8_t);
+	void setBrightness(int);
+	void setRGB(int, int, int);
 	void turnOff();
 	void runProgram(int);
